requirements.txt
beautifulsoup4==4.12.2
certifi==2023.7.22
charset-normalizer==3.2.0
comtypes==1.2.0
EasyProcess==1.1
entrypoint2==1.1
idna==3.4
MouseInfo==0.1.3
mss==9.0.1
Pillow==10.0.1
psutil==5.9.5
PyAudio==0.2.13
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
pyjokes==0.6.0
PyMsgBox==1.0.9
pyperclip==1.8.2
pypiwin32==223
PyRect==0.2.0
pyscreenshot==3.1
PyScreeze==0.1.29
pyttsx3==2.90
pytweening==1.0.7
pywin32==306
requests==2.31.0
soupsieve==2.5
SpeechRecognition==3.10.0
urllib3==2.0.5
wikipedia==1.4.0
pyautogui
pillow
pyaudio